@@ -7,43 +7,31 @@ use App\Entity\Traits\EntityTimestampableTrait;
 use App\Repository\ConnectorRepository;
 use Doctrine\ORM\Mapping as ORM;
 
-/**
- * @ORM\Entity(repositoryClass=ConnectorRepository::class)
- * @ORM\InheritanceType("SINGLE_TABLE")
- * @ORM\DiscriminatorColumn(name="discr", type="string")
- * @ORM\DiscriminatorMap({"office365" = "Office365Connector"})
- */
+#[ORM\Entity(repositoryClass: ConnectorRepository::class)]
+#[ORM\InheritanceType('SINGLE_TABLE')]
+#[ORM\DiscriminatorColumn(name: 'discr', type: 'string')]
+#[ORM\DiscriminatorMap(['office365' => 'Office365Connector'])]
 class Connector
 {
     use EntityBlameableTrait;
     use EntityTimestampableTrait;
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: 'integer')]
     private $id;
 
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
+    #[ORM\Column(type: 'string', length: 255)]
     private $name;
 
-    /**
-     * @ORM\Column(type="boolean", nullable=true)
-     */
+    #[ORM\Column(type: 'boolean', nullable: true)]
     private $active;
 
-    /**
-     * @ORM\ManyToOne(targetEntity=Domain::class, inversedBy="connectors")
-     * @ORM\JoinColumn(nullable=false, onDelete="CASCADE")
-     */
+    #[ORM\ManyToOne(targetEntity: Domain::class, inversedBy: 'connectors')]
+    #[ORM\JoinColumn(nullable: false, onDelete: 'CASCADE')]
     private $domain;
 
 
-    /**
-     * @ORM\Column(type="string", length=50)
-     */
+    #[ORM\Column(type: 'string', length: 50)]
     private $type;
 
     public function getId(): ?int
