@@ -6,33 +6,27 @@ use Doctrine\ORM\Mapping as ORM;
 
 /**
  * GroupsWblist
- *
- * @ORM\Table(name="groups_wblist")
- * @ORM\Entity
  */
+#[ORM\Table(name: 'groups_wblist')]
+#[ORM\Entity]
 class GroupsWblist
 {
 
-    /**
-    * @ORM\ManyToOne(targetEntity="App\Entity\Groups")
-    * @ORM\JoinColumn(name="group_id", nullable=true, onDelete="CASCADE")
-    * @ORM\Id
-    */
+    #[ORM\ManyToOne(targetEntity: 'App\Entity\Groups')]
+    #[ORM\JoinColumn(name: 'group_id', nullable: true, onDelete: 'CASCADE')]
+    #[ORM\Id]
     private $groups;
 
 
-    /**
-    * @ORM\ManyToOne(targetEntity="App\Entity\Mailaddr", fetch="EAGER")
-    * @ORM\JoinColumn(name="sid", nullable=true)
-    * @ORM\Id
-    */
+    #[ORM\ManyToOne(targetEntity: 'App\Entity\Mailaddr', fetch: 'EAGER')]
+    #[ORM\JoinColumn(name: 'sid', nullable: true)]
+    #[ORM\Id]
     private $mailaddr;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="wb", type="string", length=10, nullable=false)
      */
+    #[ORM\Column(name: 'wb', type: 'string', length: 10, nullable: false)]
     private $wb;
 
 
