@@ -6,9 +6,7 @@ use Doctrine\ORM\Mapping as ORM;
 use App\Entity\Traits\EntityBlameableTrait;
 use App\Entity\Traits\EntityTimestampableTrait;
 
-/**
- * @ORM\Entity(repositoryClass="App\Repository\LogRepository")
- */
+#[ORM\Entity(repositoryClass: 'App\Repository\LogRepository')]
 class Log
 {
     use EntityBlameableTrait;
@@ -16,30 +14,25 @@ class Log
 
     /**
      * @var int
-     *
-     * @ORM\Column(name="id", type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
      */
+    #[ORM\Column(name: 'id', type: 'integer')]
+    #[ORM\Id]
+    #[ORM\GeneratedValue(strategy: 'AUTO')]
     private $id;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="action", type="string", length=255, nullable=true)
      */
+    #[ORM\Column(name: 'action', type: 'string', length: 255, nullable: true)]
     private $action;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="mailId", type="string", length=255, nullable=true)
      */
+    #[ORM\Column(name: 'mailId', type: 'string', length: 255, nullable: true)]
     private $mailId;
 
-    /**
-     * @ORM\Column(type="text", nullable=true)
-     */
+    #[ORM\Column(type: 'text', nullable: true)]
     private $details;
 
     public function getId(): ?int
