@@ -6,47 +6,42 @@ use Doctrine\ORM\Mapping as ORM;
 
 /**
  * Maddr
- *
- * @ORM\Table(name="maddr", uniqueConstraints={@ORM\UniqueConstraint(name="part_email", columns={"partition_tag", "email"})})
- * @ORM\Entity
  */
+#[ORM\Table(name: 'maddr')]
+#[ORM\UniqueConstraint(name: 'part_email', columns: ['partition_tag', 'email'])]
+#[ORM\Entity]
 class Maddr
 {
     /**
      * @var int
-     *
-     * @ORM\Column(name="id", type="bigint", nullable=false, options={"unsigned"=true})
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="IDENTITY")
      */
+    #[ORM\Column(name: 'id', type: 'bigint', nullable: false, options: ['unsigned' => true])]
+    #[ORM\Id]
+    #[ORM\GeneratedValue(strategy: 'IDENTITY')]
     private $id;
 
     /**
      * @var int|null
-     *
-     * @ORM\Column(name="partition_tag", type="integer", nullable=true)
      */
+    #[ORM\Column(name: 'partition_tag', type: 'integer', nullable: true)]
     private $partitionTag = '0';
 
     /**
      * @var binary
-     *
-     * @ORM\Column(name="email", type="binary", nullable=false)
      */
+    #[ORM\Column(name: 'email', type: 'binary', nullable: false)]
     private $email;
 
     /**
      * @var int
-     *
-     * @ORM\Column(name="is_invalid", type="boolean", nullable=true)
      */
+    #[ORM\Column(name: 'is_invalid', type: 'boolean', nullable: true)]
     private $isInvalid;
 
     /**
      * @var string
-     *
-     * @ORM\Column(name="domain", type="string", length=255, nullable=false)
      */
+    #[ORM\Column(name: 'domain', type: 'string', length: 255, nullable: false)]
     private $domain;
 
     public function getId(): ?int
