@@ -6,33 +6,30 @@ use Doctrine\ORM\Mapping as ORM;
 
 /**
  * Mailaddr
- *
- * @ORM\Table(name="mailaddr", uniqueConstraints={@ORM\UniqueConstraint(name="email", columns={"email"})})
- * @ORM\Entity
  */
+#[ORM\Table(name: 'mailaddr')]
+#[ORM\UniqueConstraint(name: 'email', columns: ['email'])]
+#[ORM\Entity]
 class Mailaddr
 {
     /**
      * @var int
-     *
-     * @ORM\Column(name="id", type="integer", nullable=false, options={"unsigned"=true})
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="IDENTITY")
      */
+    #[ORM\Column(name: 'id', type: 'integer', nullable: false, options: ['unsigned' => true])]
+    #[ORM\Id]
+    #[ORM\GeneratedValue(strategy: 'IDENTITY')]
     private $id;
 
     /**
      * @var int
-     *
-     * @ORM\Column(name="priority", type="integer", nullable=false, options={"default"="7"})
      */
+    #[ORM\Column(name: 'priority', type: 'integer', nullable: false, options: ['default' => 7])]
     private $priority = '7';
 
     /**
      * @var binary
-     *
-     * @ORM\Column(name="email", type="binary", nullable=false)
      */
+    #[ORM\Column(name: 'email', type: 'binary', nullable: false)]
     private $email;
 
     public function getId(): ?int
